requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.0.6
django-bootstrap4==24.3
djangorestframework==3.15.1
pillow==10.3.0
soupsieve==2.5
sqlparse==0.5.0
